[wip] moved all system ID scripts into one file. Updated naming conventions. Updated readme
