+++ tools/system_id/README.md
@@ -14,7 +14,8 @@ This folder contains scripts to measure propellers and motors using the systemId
   * QRD1114 (RPM)
 * Calibration weights
 * M2/M3 Nylon screws
-* CF Mount (3D printed)
+* CF Mount (3D printed) TODO upload
+* Loadcell mount (3D printed, optional) TODO upload
 
 Mount CF upside down (to avoid downwash)
 
@@ -23,30 +24,10 @@ Mount CF upside down (to avoid downwash)
 ### Common
 
 1. Mount CF
-2. Switch to crazyflie-lib-python with branch `feature-emergency-stop` in order to be able to CTRL+C a script safely.
+2. Switch to crazyflie-lib-python with branch `feature-emergency-stop` in order to be able to CTRL+C a script safely. TODO how?
 3. Build firmware with configuration `make sysid_defconfig` and `make -j$(nproc)`
 4. Flash firmware by bringing the Crazyflie into flash mode and running `make cload`
 5. Run `python3 calibscale.py --uri <URI>` and follow the prompts to calibrate the load cell. This will create an output file `calibration.yaml` with the calibration data. The other scripts will read this file (other name can be specified as command line argument).
-
-### Collecting data
-
-There are multiple modes for collecting data. Each of the modes can easily be accessed by giving the correct argument to 'collect_data.py'. The combination of propellers, motors, and battery needs to be given by COMB. Since COMB can have a dash, you can use the following command in your terminal:
-
-```
-python3 collect_data.py --uri=<URI> --mode=<MODE> --comb=<COMB> 
-```
-
-The collected data is then saved accordingly to `data_<MODE>_<COMB>.csv`.
-
-In theory, you can put whatever you want for COMB or even leave it empty. We've come up with the following scheme to distinguish different setups easily:
-
-| Chars                   | Option 1    | Option 2    | Option 3      |
-| --------                | -------     | -------     | -------       |
-| 1: Propeller type       | old: -      | 2.1+: +     |
-| 2: Motor type           | 17mm: B     | 20mm: T     | Brushless: L  |
-| 3-5: Battery capacity   | 250mAh: 250 | 350mAh: 350 |
-
-As an example: The Crazyflies 2.1+ gets shipped with the new propellers, 17mm motors, and a 250mAh battery, so we would set `COMB=+B250`.
 
 ### Ramping Motors
 
@@ -56,6 +37,44 @@ This test will simply ramp the motors up and down. The results can be visualized
 python3 collect_data.py --uri <URI> --mode=ramp_motors --comb=<COMB> 
 python3 plot_data_ramp_motors.py --file=<FILENAME>
 ```
+
+### System identification
+
+The system identification is done in three steps, each called a different mode: First, the static thrust parameters are identified. Those can then be updated in the firmware and verified in the second step. Lastly, the dynamic parameters can be identified. 
+
+The scheme for all data collection and identification script is the same:
+```
+python3 collect_data.py --uri=<URI> --mode=<MODE> --comb=<COMB> 
+python3 system_id.py --mode=<MODE> --comb=<COMB> 
+```
+
+The collected data is then saved accordingly to `data_<MODE>_<COMB>.csv`.
+
+The combination of propellers, motors, and battery needs to be given by COMB. In theory, you can put whatever you want for COMB or even leave it empty. We've come up with the following scheme to distinguish different setups easily:
+
+| Chars                   | Option 1    | Option 2    | Option 3      |
+| --------                | -------     | -------     | -------       |
+| 1: Propeller type       | old: -      | 2.1+: +     |
+| 2: Motor type           | 17mm: B     | 20mm: T     | Brushless: L  |
+| 3-5: Battery capacity   | 250mAh: 250 | 350mAh: 350 |
+
+As an example: The Crazyflies 2.1+ gets shipped with the new propellers, 17mm motors, and a 250mAh battery, so we would set `COMB=+B250`.
+
+#### Static parameters
+`mode = static`
+
+In this mode, the motors are given random commands in a valid range and all the important data is stored. From that we can calculate the most important curve: Vmotors [V] -> Thrust [N], where Vmotors = PWM_CMD / PWM_MAX * Vbat. This curve helps us to calculate the needed PWM_CMD from a thrust command and the current battery voltage Vbat. Additionally, the RPM -> Thrust [N] curve and the efficiency is identified.
+
+The important parameters will be stored in params_<COMB>.yaml
+
+#### Verification
+
+TODO
+
+#### Dynamic parameters
+
+TODO
+
 
 ### System identification
